fix: migrate amount columns from bigint to decimal to avoid precision problem in database;\ fix token search api

+++ packages/tracker/src/entities/tokenInfo.entity.ts
@@ -42,7 +42,7 @@ export class TokenInfoEntity {
   @Column({ name: 'premine', nullable: true })
   premine: number;
 
-  @Column({ name: 'token_limit', type: 'bigint', nullable: true })
+  @Column({ name: 'token_limit', type: 'decimal', nullable: true })
   tokenLimit: BigInt;
 
   @Column({ name: 'deploy_block', nullable: true })
